sprint-3/step-1 feat: переделать всю авторизацию
