+++ src/pages/LoginPage/LoginPage.module.css
@@ -1,0 +1,9 @@
+.loginPage {
+  width: 100%;
+  color: hsla(240, 5%, 95%, 1);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+}
